feat(mall): complete trading and access foundations

+++ own-send-server/src/main/java/com/own/send/server/service/RdbSvc.java
@@ -11,6 +11,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
+
  
 
 @Service
@@ -31,13 +33,13 @@ public class RdbSvc{
 	 */
 	@Transactional
 	public Object save(Object obj){
-		IDomainBase objPo =  (IDomainBase)obj ;
+		IDomainBase objPo = requireDomain(obj);
  
 		if(objPo.getObjectId()!=null){
 			objPo = (IDomainBase)baseDao.find(obj,objPo.getObjectId());
-			if(objPo==null)
-				//throw new IFException("要更新的实体记录不存在！");
-				System.out.println("要更新的实体记录不存在！");
+			if (objPo == null) {
+				throw new EntityNotFoundException("Entity to update does not exist");
+			}
 			BeanUtils.copyProperties(obj,objPo);
 			baseDao.save(objPo);
 			return objPo;
@@ -53,10 +55,14 @@ public class RdbSvc{
 	 */
 	@Transactional
 	public Object update(Object obj){
-		Object objPo = baseDao.find(obj, ((IDomainBase)obj).getObjectId());
-		if(objPo==null)
-			//throw new IFException("要更新的实体记录不存在！");
-		System.out.println("要更新的实体记录不存在！");
+		IDomainBase domain = requireDomain(obj);
+		if (domain.getObjectId() == null) {
+			throw new IllegalArgumentException("Entity identifier is required for update");
+		}
+		Object objPo = baseDao.find(obj, domain.getObjectId());
+		if (objPo == null) {
+			throw new EntityNotFoundException("Entity to update does not exist");
+		}
 		BeanUtils.copyProperties(obj,objPo);
 		
 		baseDao.save(objPo);
@@ -69,14 +75,15 @@ public class RdbSvc{
 	 * @param obj
 	 */
 	public <T> void delete(Object obj){
-		if(!(obj instanceof IDomainBase))
-				//throw new IFException("该实体没有继承IDomainBase！！！");
-		System.out.println("该实体没有继承IDomainBase！！！");
-		Object key = ((IDomainBase)obj).getObjectId();
+		IDomainBase domain = requireDomain(obj);
+		Object key = domain.getObjectId();
+		if (key == null) {
+			throw new IllegalArgumentException("Entity identifier is required for delete");
+		}
 		Object object = baseDao.find(obj, key);
-		if(obj==null)
-			//throw new IFException("要删除的实体不存在或已被删除！");
-			System.out.println("要删除的实体不存在或已被删除！");
+		if (object == null) {
+			throw new EntityNotFoundException("Entity to delete does not exist");
+		}
 		baseDao.remove(object);
 	}
 	
@@ -87,10 +94,7 @@ public class RdbSvc{
 	 * @return
 	 */
 	public Object find(Object obj){
-		if(!(obj instanceof IDomainBase))
-			//throw new IFException("该实体没有继承IDomainBase！！！");
-			System.out.println("该实体没有继承IDomainBase！！！");
-		Object key = ((IDomainBase)obj).getObjectId();
+		Object key = requireDomain(obj).getObjectId();
 		Object returnO = baseDao.find(obj, key);
 		
 //		if(returnO==null)
@@ -100,9 +104,7 @@ public class RdbSvc{
 	
 	//根据实体和key进行查找
 	public Object find(Object obj,Object key){
-		if(!(obj instanceof IDomainBase))
-				//throw new IFException("该实体没有继承IDomainBase！！！");
-			System.out.println("该实体没有继承IDomainBase！！！");
+		requireDomain(obj);
   		Object returnO = baseDao.find(obj, key);
   
 		return returnO;
@@ -163,5 +165,12 @@ public class RdbSvc{
 	
 	public List<?> findList(String jsql,Map paramMap){
 		return baseDao.findList(jsql, paramMap);
+	}
+
+	private IDomainBase requireDomain(Object obj) {
+		if (!(obj instanceof IDomainBase)) {
+			throw new IllegalArgumentException("Entity must implement IDomainBase");
+		}
+		return (IDomainBase) obj;
 	}
 }
